requirements.txt
aniso8601==9.0.1
async-timeout==4.0.2
attrs==22.2.0
bearer==3.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
dbutils-gonzalo123==1.3.11
Flask==2.2.3
flask-restx==1.1.0
idna==3.4
importlib-metadata==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
marshmallow==3.19.0
packaging==23.0
psycopg2-binary==2.9.6
pyrsistent==0.19.3
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2023.3
requests==2.28.2
urllib3==1.26.15
Werkzeug==2.2.3
zipp==3.15.0
